scripts/requirements.txt
pandas>=2.2.0
psycopg[binary]>=3.1.0
sqlalchemy>=2.0.25
python-dotenv>=1.0.1 